+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix stretched avatars in channel list
 - Fixed file permssions on Linux
 - Reworked file/folder/remote URL fetching dialogs
+- Fixed "jumpy" scroll bars while changing channels or messages
 
 ## [0.0.2] - 2020-07-31
 ### Added